+++ docs/source/docs_index.rst
@@ -19,6 +19,5 @@ Flytekit API Reference
    tasks.extend
    types.extend
    experimental
-   inference
    pyflyte
    contributing
