--- docs/source/inference.rst
@@ -1,4 +1,0 @@
-.. automodule:: flytekit.core.inference
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
