+++ docs/source/plugins/inference.rst
@@ -1,0 +1,12 @@
+.. _inference:
+
+#########################
+Model Inference reference
+#########################
+
+.. tags:: Integration, Serving, Inference
+
+.. automodule:: flytekitplugins.inference
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
