+++ plugins/flytekit-inference/flytekitplugins/inference/__init__.py
@@ -1,0 +1,13 @@
+"""
+.. currentmodule:: flytekitplugins.inference
+
+.. autosummary::
+   :nosignatures:
+   :template: custom.rst
+   :toctree: generated/
+
+   NIM
+   NIMSecrets
+"""
+
+from .nim.serve import NIM, NIMSecrets
