+++ plugins/flytekit-inference/flytekitplugins/inference/nim/serve.py
@@ -1,24 +1,7 @@
-"""
-=========
-Inference
-=========
-
-.. currentmodule:: flytekit.core.inference
-
-This module includes inference subclasses that extend the `ModelInferenceTemplate`.
-
-.. autosummary::
-   :nosignatures:
-   :template: custom.rst
-   :toctree: generated/
-
-   NIM
-"""
-
 from dataclasses import dataclass
 from typing import Optional
 
-from .utils import ModelInferenceTemplate
+from ..sidecar_template import ModelInferenceTemplate
 
 
 @dataclass
